change buffer lengths of sysctl(2) to be size_t's
SCCS-vsn: 7.20

+++ sys/vm/vm_meter.c
@@ -30,7 +30,7 @@
  * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  *
- *	@(#)vm_meter.c	7.19 (Berkeley) 03/02/93
+ *	@(#)vm_meter.c	7.20 (Berkeley) 03/04/93
  */
 
 #include <sys/param.h>
@@ -105,9 +105,9 @@ vm_sysctl(name, namelen, oldp, oldlenp, newp, newlen, p)
 	int *name;
 	u_int namelen;
 	void *oldp;
-	u_int *oldlenp;
+	size_t *oldlenp;
 	void *newp;
-	u_int newlen;
+	size_t newlen;
 	struct proc *p;
 {
 	struct vmtotal vmtotals;
